Refactor file I/O and formatting utilities
This commit restructures utility code into more focused modules:

- Move common I/O operations into new `io.rs` module - Create dedicated
`format.rs` module for text formatting - Rename `output.rs` to `log.rs`
for logging utilities - Add documentation and examples for each module -
Add unit tests for new functionality

The changes improve code organization and make the utilities more
maintainable without changing their behavior.
