+++ crates/bargo-core/src/util/log.rs
@@ -1,18 +1,43 @@
-use std::path::Path;
-
-use super::timer::Timer;
+//! Logging and output formatting utilities for bargo
+//!
+//! This module provides utilities for colored terminal output, logging messages,
+//! and formatted banners used throughout the bargo workflow.
+//!
+//! ## Key Features
+//!
+//! - Colored terminal output with automatic NO_COLOR support
+//! - Success, info, and error message formatting
+//! - ASCII art banners for different operations
+//! - TTY detection for proper color handling
+//!
+//! ## Examples
+//!
+//! ```ignore
+//! use bargo_core::util::log::{success, info, print_banner};
+//!
+//! // Print colored success message
+//! println!("{}", success("Build completed successfully"));
+//!
+//! // Print info message
+//! println!("{}", info("Starting verification process"));
+//!
+//! // Print operation banner
+//! print_banner("build");
+//! ```
 
 /// ANSI color codes for terminal output
 pub mod colors {
     pub const RESET: &str = "\x1b[0m";
     pub const BOLD: &str = "\x1b[1m";
     pub const GREEN: &str = "\x1b[32m";
-
     pub const GRAY: &str = "\x1b[90m";
     pub const BRIGHT_GREEN: &str = "\x1b[92m";
     pub const BRIGHT_BLUE: &str = "\x1b[94m";
     pub const BRIGHT_CYAN: &str = "\x1b[96m";
 }
+
+// Placeholder functions - these will be moved here from other modules
+// in Checkpoint B
 
 /// Format text with color
 pub fn colorize(text: &str, color: &str) -> String {
@@ -33,12 +58,7 @@ pub fn info(text: &str) -> String {
     colorize(&format!("ℹ️ {}", text), colors::BRIGHT_BLUE)
 }
 
-/// Create path text with cyan color
-pub fn path(text: &str) -> String {
-    colorize(text, colors::BRIGHT_CYAN)
-}
-
-/// ASCII art banners for different operations
+/// Print ASCII art banners for different operations
 pub fn print_banner(operation: &str) {
     let banner = match operation {
         "build" => {
@@ -81,32 +101,36 @@ pub fn print_banner(operation: &str) {
     println!("{}", colorize(banner, colors::BRIGHT_BLUE));
 }
 
-/// Format file size in human-readable format
-pub fn format_file_size(path: &Path) -> String {
-    match std::fs::metadata(path) {
-        Ok(metadata) => {
-            let size = metadata.len();
-            if size < 1024 {
-                format!("{} B", size)
-            } else if size < 1024 * 1024 {
-                format!("{:.1} KB", size as f64 / 1024.0)
-            } else {
-                format!("{:.1} MB", size as f64 / (1024.0 * 1024.0))
-            }
-        }
-        Err(_) => "unknown size".to_string(),
-    }
-}
+#[cfg(test)]
+mod tests {
+    use super::*;
 
-/// Format operation result with file size and timing
-pub fn format_operation_result(operation: &str, file_path: &Path, timer: &Timer) -> String {
-    let size = format_file_size(file_path);
-    let elapsed = timer.elapsed();
-    format!(
-        "{} → {} ({}, {})",
-        operation,
-        file_path.display(),
-        size,
-        elapsed
-    )
+    #[test]
+    fn test_success() {
+        let result = success("test message");
+        assert!(result.contains("✅"));
+        assert!(result.contains("test message"));
+    }
+
+    #[test]
+    fn test_info() {
+        let result = info("test info");
+        assert!(result.contains("ℹ️"));
+        assert!(result.contains("test info"));
+    }
+
+    #[test]
+    fn test_colorize() {
+        let result = colorize("test", colors::GREEN);
+        // Should contain the text (color codes may vary based on environment)
+        assert!(result.contains("test"));
+    }
+
+    #[test]
+    fn test_print_banner() {
+        // This test just ensures the function doesn't panic
+        print_banner("build");
+        print_banner("prove");
+        print_banner("unknown_operation");
+    }
 }
